twitter-bot/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
tweepy==4.14.0
google-generativeai==0.7.2
apscheduler==3.10.4
python-dotenv==1.0.1
sqlmodel==0.0.19
python-multipart==0.0.9
